refactor(launcher): move epoch helper into artemis_cli.utils subpackage
